% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send-sms.R
\name{send_sms}
\alias{send_sms}
\title{Send SMS.}
\usage{
send_sms(to, message)
}
\arguments{
\item{to}{Number to send SMS to.}

\item{message}{SMS message.}

\item{from}{Number to send SMS from.}
}
\value{
GET url response.
}
\description{
Send SMS.
}
\examples{
send_sms("2783000000", "2783000001", "Hello world.")
send_sms()
}
