% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsportal-gentoken.R
\name{smsportal_gentoken}
\alias{smsportal_gentoken}
\title{Generate SMSPortal API token.}
\usage{
smsportal_gentoken(client_id, secret_key)
}
\arguments{
\item{client_id}{SMSPortal client ID.}

\item{secret_key}{SMSPortal secret key.}
}
\value{
Token.
}
\description{
Generate SMSPortal API token.
}
\examples{
smsportal_token(client_id = "", secret_key = "")
smsportal_token()
}
