% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsportal-token.R
\name{smsportal_token}
\alias{smsportal_token}
\title{Set or query SMSPortal API token.}
\usage{
smsportal_token(token = NULL)
}
\arguments{
\item{token}{SMSPortal API token.}
}
\value{
Token.
}
\description{
Set or query SMSPortal API token.
}
\examples{
smsportal_token("eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJsb2dpbklkIjoiMzE2NDQ1IiwiaXNzIjoiU21zUG9ydGFsU2VjdXJpdHlBcGkiLCJhdWQiOiJBbGwiLCJleHAiOjE1Nzg3NDE2NDcsIm5iZiI6MTU3ODY1NTI0N30.uul-X3uQp3Nyn7in_5CFcZCHi3_M2a151jOzeexuFCo")
smsportal_token()
}
